test(precision-recall): Move tests from file to folder
