+++ skore/tests/unit/sklearn/plot/precision_recall_curve/test_cross_validation.py
@@ -1,0 +1,147 @@
+import matplotlib as mpl
+import numpy as np
+import pytest
+from skore import CrossValidationReport
+from skore.sklearn._plot import PrecisionRecallCurveDisplay
+from skore.sklearn._plot.utils import sample_mpl_colormap
+
+
+@pytest.mark.parametrize("data_source", ["train", "test", "X_y"])
+def test_precision_recall_curve_cross_validation_display_binary_classification(
+    pyplot, binary_classification_data_no_split, data_source
+):
+    """Check the attributes and default plotting behaviour of the
+    precision-recall curve plot with binary data.
+    """
+    (estimator, X, y), cv = binary_classification_data_no_split, 3
+    report = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv)
+    if data_source == "X_y":
+        precision_recall_kwargs = {"data_source": data_source, "X": X, "y": y}
+    else:
+        precision_recall_kwargs = {"data_source": data_source}
+
+    display = report.metrics.precision_recall(**precision_recall_kwargs)
+    assert isinstance(display, PrecisionRecallCurveDisplay)
+
+    # check the structure of the attributes
+    pos_label = report.estimator_reports_[0].estimator_.classes_[1]
+    for attr_name in ("precision", "recall", "average_precision"):
+        assert isinstance(getattr(display, attr_name), dict)
+        assert len(getattr(display, attr_name)) == 1
+
+        attr = getattr(display, attr_name)
+        assert list(attr.keys()) == [pos_label]
+        assert list(attr.keys()) == [display.pos_label]
+        assert isinstance(attr[pos_label], list)
+        assert len(attr[pos_label]) == cv
+
+    display.plot()
+    assert hasattr(display, "ax_")
+    assert hasattr(display, "figure_")
+    assert isinstance(display.lines_, list)
+    assert len(display.lines_) == cv
+    expected_colors = sample_mpl_colormap(pyplot.cm.tab10, 10)
+    for split_idx, line in enumerate(display.lines_):
+        assert isinstance(line, mpl.lines.Line2D)
+        assert line.get_label() == (
+            f"Estimator of fold #{split_idx + 1} "
+            f"(AP = {display.average_precision[pos_label][split_idx]:0.2f})"
+        )
+        assert mpl.colors.to_rgba(line.get_color()) == expected_colors[split_idx]
+
+    assert isinstance(display.ax_, mpl.axes.Axes)
+    legend = display.ax_.get_legend()
+    data_source_title = "external" if data_source == "X_y" else data_source
+    assert (
+        legend.get_title().get_text()
+        == f"LogisticRegression on $\\bf{{{data_source_title}}}$ set"
+    )
+    assert len(legend.get_texts()) == 3
+
+    assert display.ax_.get_xlabel() == "Recall\n(Positive label: 1)"
+    assert display.ax_.get_ylabel() == "Precision\n(Positive label: 1)"
+    assert display.ax_.get_adjustable() == "box"
+    assert display.ax_.get_aspect() in ("equal", 1.0)
+    assert display.ax_.get_xlim() == display.ax_.get_ylim() == (-0.01, 1.01)
+
+
+@pytest.mark.parametrize("data_source", ["train", "test", "X_y"])
+def test_precision_recall_curve_cross_validation_display_multiclass_classification(
+    pyplot, multiclass_classification_data_no_split, data_source
+):
+    """Check the attributes and default plotting behaviour of the precision-recall
+    curve plot with multiclass data.
+    """
+    (estimator, X, y), cv = multiclass_classification_data_no_split, 3
+    report = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv)
+    if data_source == "X_y":
+        precision_recall_kwargs = {"data_source": data_source, "X": X, "y": y}
+    else:
+        precision_recall_kwargs = {"data_source": data_source}
+
+    display = report.metrics.precision_recall(**precision_recall_kwargs)
+    assert isinstance(display, PrecisionRecallCurveDisplay)
+
+    # check the structure of the attributes
+    class_labels = report.estimator_reports_[0].estimator_.classes_
+    for attr_name in ("precision", "recall", "average_precision"):
+        assert isinstance(getattr(display, attr_name), dict)
+        assert len(getattr(display, attr_name)) == len(class_labels)
+
+        attr = getattr(display, attr_name)
+        for class_label in class_labels:
+            assert isinstance(attr[class_label], list)
+            assert len(attr[class_label]) == cv
+
+    display.plot()
+    assert isinstance(display.lines_, list)
+    assert len(display.lines_) == len(class_labels) * cv
+    default_colors = sample_mpl_colormap(pyplot.cm.tab10, 10)
+    for class_label, expected_color in zip(class_labels, default_colors):
+        for split_idx in range(cv):
+            precision_recall_curve_mpl = display.lines_[class_label * cv + split_idx]
+            assert isinstance(precision_recall_curve_mpl, mpl.lines.Line2D)
+            if split_idx == 0:
+                assert precision_recall_curve_mpl.get_label() == (
+                    f"{str(class_label).title()} "
+                    f"(AP = {np.mean(display.average_precision[class_label]):0.2f}"
+                    f" +/- {np.std(display.average_precision[class_label]):0.2f})"
+                )
+            assert precision_recall_curve_mpl.get_color() == expected_color
+
+    assert isinstance(display.ax_, mpl.axes.Axes)
+    legend = display.ax_.get_legend()
+    data_source_title = "external" if data_source == "X_y" else data_source
+    assert (
+        legend.get_title().get_text()
+        == f"LogisticRegression on $\\bf{{{data_source_title}}}$ set"
+    )
+    assert len(legend.get_texts()) == 3
+
+    assert display.ax_.get_xlabel() == "Recall"
+    assert display.ax_.get_ylabel() == "Precision"
+    assert display.ax_.get_adjustable() == "box"
+    assert display.ax_.get_aspect() in ("equal", 1.0)
+    assert display.ax_.get_xlim() == display.ax_.get_ylim() == (-0.01, 1.01)
+
+
+@pytest.mark.parametrize(
+    "fixture_name",
+    ["binary_classification_data_no_split", "multiclass_classification_data_no_split"],
+)
+@pytest.mark.parametrize("pr_curve_kwargs", [[{"color": "red"}], "unknown"])
+def test_pr_curve_display_cross_validation_multiple_roc_curve_kwargs_error(
+    pyplot, fixture_name, request, pr_curve_kwargs
+):
+    """Check that we raise a proper error message when passing an inappropriate
+    value for the `pr_curve_kwargs` argument."""
+    (estimator, X, y), cv = request.getfixturevalue(fixture_name), 3
+
+    report = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv)
+    display = report.metrics.precision_recall()
+    err_msg = (
+        "You intend to plot multiple curves. We expect `pr_curve_kwargs` to be a list "
+        "of dictionaries"
+    )
+    with pytest.raises(ValueError, match=err_msg):
+        display.plot(pr_curve_kwargs=pr_curve_kwargs)
